Users can login/logout
Add links controller
Add method for current user in application controller
Add sessions for logging in and out
Add validations on users.
Add navbar
Add new session view
And a showpage for users
Add routes for logging in and out
Add tests for users to sign up, login and logout.